requirements.txt
openai-whisper        
gpt4all               
pyttsx3               
sounddevice           
numpy                 
